==> 1_prepare_data.ipynb <==
# 1. Download and prepare data

Data concerns reviews of books on Amazon and was downloaded from<br>
https://nijianmo.github.io/amazon/index.html

Due to the size of the dataset, it was sometimes necessary to use generators in preprocessing.

For a bigger contrast in sentiment, only reviews with an overall rating of 1.3 or 5 were selected. As 5-rated reviews made up nearly 80% of the collection, it was decided to undersampling to balance the grades. Then the missing data and duplicates were removed. In the next step, the target variable was created from the `overall` variable.

Then, the texts of the reviews were cleaned up. Removed:
* URL
* newline characters
* special signs
* numbers
* tags

Then, using `gensim.utils.simple_preprocess()`, further purification of the texts and tokenization were performed. Stopwords were removed using the `spacy` module and the `en_core_web_md` language model. There were used `pad_sequence` for padding - to prepare strings of equal length for each review.<br>
Finally, the dataset was splitted into train (75%) and test (25%).

# Dowloade language model from spacy
!python -m spacy download en_core_web_md

# NOTE!!! MUST restart runtime

# Download the data - amazon books reviews
!wget http://snap.stanford.edu/data/amazon/productGraph/categoryFiles/reviews_Books_5.json.gz

from google.colab import drive
drive.mount('/content/drive')

import gzip
import re
import string
import numpy as np
import pandas as pd
import pickle

import gensim
from gensim.utils import simple_preprocess
import spacy
from nltk.tokenize.treebank import TreebankWordDetokenizer

import tensorflow as tf
import keras
from keras.preprocessing.text import Tokenizer
from keras.preprocessing.sequence import pad_sequences
from sklearn.model_selection import train_test_split

def parse(load_path, filter = True):
    """
    The generator parses compressed .json.gz file,
    extracts two values: 'overall' and 'reviewText',
    selects only these reviews with a rating of 1.3 or 5,
    and finally yields the given observation.

    Parameters:
    load_path - path to compressed .json.gz directory
    """
    g = gzip.open(load_path, 'rb')
    for l in g:
        l = eval(l)
        l = {key: l[key] for key in l.keys() & {'overall','reviewText'}}
        if filter:
            if l['overall'] in [2,4]:
                pass
            else:
                yield l
        else:
          yield l

def parse_to_df(load_path, filter):
    """
    The function parses the path load_path line by line
    and saves reviews to DataFrame,

    Parameters:
    load_path - path to compressed .json.gz directory
    """
    i = 0
    df = {}
    for d in parse(load_path, filter):
        df[i] = d
        i += 1
    return pd.DataFrame.from_dict(df, orient='index')

df = parse_to_df('reviews_Books_5.json.gz', filter=True)
df

# How many reviews are in each class?
df['overall'].value_counts()

# Target is unbalanced. Class '5' is more represented than others.
# Data are shufled so we will take first 165.000 of reviews of each class instead of using undersampling.
n_ = 165000
print(n_)

df = pd.concat([
    df[df['overall'] == 5][:n_],
    df[df['overall'] == 3][:n_],
    df[df['overall'] == 1][:n_]
])

# Check if there are missing values
df.isnull().sum()

# Check if there are duplicated reviews, if so remove them
print('Number of duplicates:', df.duplicated().sum())

df.drop_duplicates(inplace=True)

# Create target varaible - one hot encode 'overall'
# 0 - negative
# 1 - neutral
# 2 - positive

labels_matrix = np.array(df['overall'])
y = []
for i in range(len(labels_matrix)):
    if labels_matrix[i] == 1:
        y.append(0)
    if labels_matrix[i] == 3:
        y.append(1)
    if labels_matrix[i] == 5:
        y.append(2)
y = np.array(y)
labels_matrix = tf.keras.utils.to_categorical(y, 3, dtype="int32")
del y

# Function to clean text data
def clean_data(data):
    url_pattern = re.compile(r'https?://\S+|www\.\S+')
    data = url_pattern.sub(r'', data)
    data = re.sub('\n', ' ', data)
    data = re.sub("\'", "", data)
    data = re.sub('\$[a-zA-Z0-9]*', ' ', data)
    data = re.sub('\@[a-zA-Z0-9]*', ' ', data)
    data = re.sub('\d+', ' ', data)
    data = re.sub(r'[^\w]', ' ', data)
    return data



data_to_list = []
temp = df['reviewText'].values.tolist()
for i in range(len(temp)):
    data_to_list.append(clean_data(temp[i]))

del temp
del df
list(data_to_list[:5])

# Conversion doc for lists of tokens ['I like the item'] --> [['I','like','the','item']]
def sent_to_words(data_to_list):
    for sentence in data_to_list:
        yield gensim.utils.simple_preprocess(str(sentence), deacc=True)

with open('after_tokenization.txt', 'w') as f:
    for line in sent_to_words(data_to_list):
        f.write("%s\n" % line)

del data_to_list

!ls -lh
!head -5 after_tokenization.txt

with open('after_tokenization.txt', 'r') as f:
    data_words = []
    for elem in f:
        data_words.append(eval(elem))

print(*data_words[:10], sep='\n')

# load eng language model from spacy - medium
nlp = spacy.load('en_core_web_md')

# create stopwords list with spacy 
stopwordlist = nlp.Defaults.stop_words

# remove stopwords
for i in range(len(data_words)):
  data_words[i] = [word for word in data_words[i] if word not in stopwordlist]

print(*data_words[:10], sep='\n')

# detokenization
def detokenize(text):
    return TreebankWordDetokenizer().detokenize(text)


data = []
for i in range(len(data_words)):
    data.append(detokenize(data_words[i]))

del data_words
print(*data[:10], sep='\n')

# tokenize and padding to max_len=600 with keras
max_words = 5000
max_len = 600

tokenizer = Tokenizer(num_words=max_words)
tokenizer.fit_on_texts(data)
sequences = tokenizer.texts_to_sequences(data)
reviews = pad_sequences(sequences, maxlen=max_len)
del data
del sequences
print(reviews)

print(labels_matrix)

# train-test split
X_train, X_test, y_train, y_test = train_test_split(reviews, labels_matrix, random_state=24, shuffle=True)

print('TRAIN:', len(y_train))
print('TRAIN:', len(y_test))
del reviews

# Save splitted data for modeling
with open('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/X_train.npy', 'wb') as f:
    np.save(f, X_train)

with open('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/X_test.npy', 'wb') as f:
    np.save(f, X_test)

with open('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/y_train.npy', 'wb') as f:
    np.save(f, y_train)

with open('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/y_test.npy', 'wb') as f:
    np.save(f, y_test)

# save tokenizer to file with pickle
with open('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/tokenizer.pickle', 'wb') as f:
    pickle.dump(tokenizer, f, protocol=pickle.HIGHEST_PROTOCOL)
