==> 3_training_BILSTM.ipynb <==
# 3. Training BILSTM

BILSTM network training took place over many sessions due to the long computation time. We trained only 12 epoch.

from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import pickle

import tensorflow as tf
import keras
from keras.models import Sequential
from keras import layers
from tensorflow.keras.optimizers import RMSprop, Adam

# Read the data
X_train = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/X_train.npy')
X_test = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/X_test.npy')
y_train = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/y_train.npy')
y_test = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/y_test.npy')

# read pretrained BILSTM model 
model_BILSTM = tf.keras.models.load_model('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_1')

# Train 
history = model_BILSTM.fit(X_train, 
                     y_train, 
                     epochs=2,
                     validation_data=(X_test, y_test),
                     verbose = 1)

# Save the model
model_BILSTM.save('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_3')

# Train 
history = model_BILSTM.fit(X_train, 
                     y_train, 
                     epochs=2,
                     validation_data=(X_test, y_test),
                     verbose = 1)

# Save the model
model_BILSTM.save('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_5')

# Train 
history = model_BILSTM.fit(X_train, 
                     y_train, 
                     epochs=2,
                     validation_data=(X_test, y_test),
                     verbose = 1)

# Save the model
model_BILSTM.save('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_7')

# Train 
history = model_BILSTM.fit(X_train, 
                     y_train, 
                     epochs=2,
                     validation_data=(X_test, y_test),
                     verbose = 1)

# Save the model
model_BILSTM.save('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_9')

# Train 
history = model_BILSTM.fit(X_train, 
                     y_train, 
                     epochs=1,
                     validation_data=(X_test, y_test),
                     verbose = 1)

# Save the model
model_BILSTM.save('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_10')

# Train 
history = model_BILSTM.fit(X_train, 
                     y_train, 
                     epochs=2,
                     validation_data=(X_test, y_test),
                     verbose = 1)

# Save the model
model_BILSTM.save('/content/drive/MyDrive/INL_PROJEKT/BILSTM/model_epoch_12')
