==> 4_init_CONV.ipynb <==
# 3. Initialization of CNN model

To classify the sentiment of a given review, we will also use the Convolutional Neural Network (CNN) model. The neural networks were created using `tensorflow`.

For the data prepared in the script *1_prepare_data*, we initiated a CNN network with two convolution layers. Then trained one epoch - 2 minuts on GPU with Google Colab.

from google.colab import drive
drive.mount('/content/drive')

import numpy as np
import pandas as pd
import pickle

import tensorflow as tf
import keras
from keras.models import Sequential
from keras import layers
from tensorflow.keras.optimizers import RMSprop, Adam
from keras import regularizers

# Read the data
X_train = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/X_train.npy')
X_test = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/X_test.npy')
y_train = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/y_train.npy')
y_test = np.load('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/y_test.npy')

# Read tokenizer
with open('/content/drive/MyDrive/INL_PROJEKT/preprocessed_data/tokenizer.pickle', 'rb') as f:
    tokenizer = pickle.load(f)

max_words = 5000
max_len = 600

model_CONV = Sequential()
model_CONV.add(layers.Embedding(max_words, 30, input_length=max_len))
model_CONV.add(layers.Conv1D(20, 6, activation='relu',kernel_regularizer=regularizers.l1_l2(l1=0.001, l2=0.001),bias_regularizer=regularizers.l2(0.001)))
model_CONV.add(layers.MaxPooling1D(5))
model_CONV.add(layers.Conv1D(20, 6, activation='relu',kernel_regularizer=regularizers.l1_l2(l1=0.0006, l2=0.0005),bias_regularizer=regularizers.l2(0.0004)))
model_CONV.add(layers.GlobalMaxPooling1D())
model_CONV.add(layers.Dense(3,activation='softmax'))
model_CONV.compile(optimizer='rmsprop',loss='categorical_crossentropy',metrics=['acc'])

# Architecture of model
model_CONV.summary()

# Training CNN - one epoch
history = model_CONV.fit(X_train, 
                         y_train,
                         epochs=1,
                         validation_data=(X_test, y_test),
                         verbose=1)

# Save the model
model_CONV.save('/content/drive/MyDrive/INL_PROJEKT/CONV/model_epoch_1')
